Add more changes lost during final rebase

+++ src/sched/lthread.c
@@ -613,7 +613,7 @@ int lthread_create_primitive(
     lt->itlssz = libc.tls_size;
     if (libc.tls_size)
     {
-        if ((lt->itls = (uint8_t*)enclave_mmap(
+        if ((intptr_t)(lt->itls = (uint8_t*)enclave_mmap(
                  0,
                  lt->itlssz,
                  0, /* map_fixed */
@@ -707,7 +707,7 @@ int lthread_create(
         return -1;
     }
     lt->attr.stack = attrp ? attrp->stack : 0;
-    if ((!lt->attr.stack) && ((lt->attr.stack = enclave_mmap(
+    if ((!lt->attr.stack) && ((intptr_t)(lt->attr.stack = enclave_mmap(
                                    0,
                                    stack_size,
                                    0, /* map_fixed */
@@ -723,7 +723,7 @@ int lthread_create(
     lt->itlssz = libc.tls_size;
     if (libc.tls_size)
     {
-        if ((lt->itls = (uint8_t*)enclave_mmap(
+        if ((intptr_t)(lt->itls = (uint8_t*)enclave_mmap(
                  0,
                  lt->itlssz,
                  0, /* map_fixed */
